+++ src/main/resources/META-INF/plugin.xml
@@ -36,7 +36,8 @@
                            implementationClass="com.github.korniloval.matlab.highlighting.MatlabPairedBraceMatcher"/>
         <quoteHandler fileType="Matlab" className="com.github.korniloval.matlab.editor.actions.MatlabQuoteHandler"/>
 
-        <completion.contributor language="Matlab" implementationClass="com.github.korniloval.matlab.completion.MatlabCompletionContributor"/>
+        <completion.contributor language="Matlab"
+                                implementationClass="com.github.korniloval.matlab.completion.MatlabKeywordCompletionContributor"/>
         <lang.findUsagesProvider language="Matlab"
                                  implementationClass="com.github.korniloval.matlab.MatlabFindUsagesProvider"/>
     </extensions>
