+++ src/test/resources/completion/If.m
@@ -1,0 +1,3 @@
+if
+  <caret>
+end
