+++ src/test/resources/completion/WhileCondition.m
@@ -1,0 +1,3 @@
+while <caret>
+
+end
